services/task_service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-dotenv
redis[async]>=4.6.0